Inline sync dialog logic in SyncActivity
